Fixing warnings (patch by Alfred E. Heggestad)
git-svn-id: http://svn.xiph.org/trunk/speex@9972 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/speex_header.c
@@ -138,7 +138,7 @@ SpeexHeader *speex_packet_to_header(char *packet, int size)
       }
    
    /*FIXME: Do we allow larger headers?*/
-   if (size < sizeof(SpeexHeader))
+   if (size < (int)sizeof(SpeexHeader))
    {
       speex_warning("Speex header too small");
       return NULL;
